+++ PayPalNativePayments/src/main/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutRequest.kt
@@ -7,5 +7,11 @@ data class PayPalNativeCheckoutRequest(
     /**
      * The order ID associated with the request.
      */
-    val orderId: String
+    val orderId: String,
+
+    /**
+     * Optional: User email to initiate a quicker authentication flow
+     * in cases where the user has a PayPal Account with the same email.
+     */
+    val userAuthenticationEmail: String? = null
 )